requirements.txt
strands-agents
'strands-agents[ollama]'
strands-agents-tools
